+++ pysat/tests/test_constellation.py
@@ -6,7 +6,6 @@
 """Unit tests for the `constellation` class and methods."""
 
 import datetime as dt
-from io import StringIO
 import logging
 import pandas as pds
 import pytest
@@ -51,29 +50,24 @@ class TestConstellationInitReg(TestWithRegistration):
         assert str(verr).find("no registered packages match input") >= 0
         return
 
-    def test_some_bad_construct_constellation(self):
+    def test_some_bad_construct_constellation(self, caplog):
         """Test partial load and log warning if some inputs are unregistered."""
-
-        # Initialize logging
-        log_capture = StringIO()
-        pysat.logger.addHandler(logging.StreamHandler(log_capture))
-        pysat.logger.setLevel(logging.WARNING)
 
         # Register fake Instrument modules
         pysat.utils.registry.register(self.module_names)
 
         # Load the Constellation and capture log output
-        const = pysat.Constellation(platforms=['Executor', 'platname1'],
-                                    tags=[''])
-        log_out = log_capture.getvalue()
+        with caplog.at_level(logging.WARNING, logger='pysat'):
+            const = pysat.Constellation(platforms=['Executor', 'platname1'],
+                                        tags=[''])
 
         # Test the partial Constellation initialization
         assert len(const.instruments) == 2
 
         # Test the log warning
-        assert log_out.find("unable to load some platforms") >= 0
+        assert caplog.text.find("unable to load some platforms") >= 0
 
-        del log_capture, log_out, const
+        del const
         return